backend/requirements.txt
fastapi
uvicorn
google-generativeai
apscheduler
python-dotenv
supabase
